# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Crew Challenges API
  version: 1.0.0
paths:
...
  /api/v1/crew:
    get:
      operationId: crew_retrieve
      description: The crew the user is acting in, with its members in the order they
        joined.
      tags:
      - crew
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewDetailOut'
          description: ''
  /api/v1/crew/invites:
    get:
      operationId: crew_invites_list
      description: 'Admin: the crew''s pending invites, and creating a single-use
        invite valid for 7 days.'
      tags:
      - crew
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingInviteOut'
          description: ''
    post:
      operationId: crew_invites_create
      description: 'Admin: the crew''s pending invites, and creating a single-use
        invite valid for 7 days.'
      tags:
      - crew
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteOut'
          description: ''
  /api/v1/crew/invites/{invite_id}:
    delete:
      operationId: crew_invites_destroy
      description: 'Admin: cancel an invite nobody has used yet.'
      parameters:
      - in: path
        name: invite_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - crew
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/invites/{code}:
    get:
      operationId: invites_retrieve
      description: 'Public: what the join page shows before the person signs up.'
      parameters:
      - in: path
        name: code
        schema:
          type: string
        required: true
      tags:
      - invites
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitePreviewOut'
          description: ''
  /api/v1/invites/{code}/accept:
    post:
      operationId: invites_accept
      description: 'Public: create an account from an invite, join the crew, and log
        in.'
      parameters:
      - in: path
        name: code
        schema:
          type: string
        required: true
      tags:
      - invites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteInRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
          description: ''
  /api/v1/invites/{code}/join:
    post:
      operationId: invites_join
      description: 'Logged in: join the invite''s crew with the account you already
        have.'
      parameters:
      - in: path
        name: code
        schema:
          type: string
        required: true
      tags:
      - invites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinWithAccountInRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
          description: ''
  /api/v1/me:
    get:
      operationId: me_retrieve
      description: The logged-in user, their membership and crew (member and crew
        are null outside a crew).
      tags:
      - me
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
          description: ''
    patch:
      operationId: me_update
      description: The logged-in user, their membership and crew (member and crew
        are null outside a crew).
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMePatchInRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
          description: ''
  /api/v1/me/crew:
    put:
      operationId: me_crew_update
      description: Switch the crew the user acts in (for people in several crews).
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveCrewInRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
          description: ''
...
components:
  schemas:
    AcceptInviteInRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 30
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 128
        display_name:
          type: string
          minLength: 1
          maxLength: 40
        preferred_language:
          allOf:
          - $ref: '#/components/schemas/PreferredLanguageEnum'
          description: |-
            The language used to sign up.

            * `ro` - Romana
            * `en` - English
      required:
      - display_name
      - password
      - username
    ActiveCrewInRequest:
      type: object
      properties:
        crew_id:
          type: string
          format: uuid
      required:
      - crew_id
...
    CrewDetailOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberOut'
          description: In the order they joined.
      required:
      - id
      - members
      - name
      - timezone
    CrewOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
      required:
      - id
      - name
      - timezone
...
    InviteOut:
      type: object
      properties:
        code:
          type: string
        url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
      required:
      - code
      - expires_at
      - url
    InvitePreviewOut:
      type: object
      properties:
        crew_name:
          type: string
        status:
          $ref: '#/components/schemas/StatusEnum'
        expires_at:
          type: string
          format: date-time
        invited_by:
          allOf:
          - $ref: '#/components/schemas/MemberSummaryOut'
          nullable: true
          description: Who created the invite. Only for valid invites.
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberSummaryOut'
          description: Members in the order they joined. Empty unless the invite is
            valid.
        already_member:
          type: boolean
          description: True when the logged-in user is already in this crew.
        crew_id:
          type: string
          format: uuid
          nullable: true
          description: The crew's id, only for its own members (to switch to it).
      required:
      - already_member
      - crew_id
      - crew_name
      - expires_at
      - invited_by
      - members
      - status
    JoinWithAccountInRequest:
      type: object
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 40
      required:
      - display_name
...
    MeOut:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        member:
          allOf:
          - $ref: '#/components/schemas/MemberOut'
          nullable: true
        crew:
          allOf:
          - $ref: '#/components/schemas/CrewOut'
          nullable: true
        crews:
          type: array
          items:
            $ref: '#/components/schemas/MembershipOut'
          description: Every crew the user belongs to, by name.
      required:
      - crew
      - crews
      - member
      - user
...
    MemberOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        role:
          type: string
        avatar_seed:
          type: string
      required:
      - avatar_seed
      - display_name
      - id
      - role
...
    MemberSummaryOut:
      type: object
      properties:
        display_name:
          type: string
        avatar_seed:
          type: string
      required:
      - avatar_seed
      - display_name
    MembershipOut:
      type: object
      properties:
        crew_id:
          type: string
          format: uuid
        crew_name:
          type: string
        display_name:
          type: string
        role:
          type: string
      required:
      - crew_id
      - crew_name
      - display_name
      - role
...
    PatchedMePatchInRequest:
      type: object
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 40
        preferred_language:
          $ref: '#/components/schemas/PreferredLanguageEnum'
    PendingInviteOut:
      type: object
      properties:
        code:
          type: string
        url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        created_by:
          allOf:
          - $ref: '#/components/schemas/MemberSummaryOut'
          nullable: true
      required:
      - code
      - created_by
      - expires_at
      - id
      - url
...
    PreferredLanguageEnum:
      enum:
      - ro
      - en
      type: string
      description: |-
        * `ro` - Romana
        * `en` - English
...
    StatusEnum:
      enum:
      - valid
      - expired
      - used
      type: string
      description: |-
        * `valid` - valid
        * `expired` - expired
        * `used` - used
...
    UserOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        preferred_language:
          $ref: '#/components/schemas/PreferredLanguageEnum'
      required:
      - id
      - preferred_language
      - username
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: crew_session